feat: delete last product from the cart
